requirements.txt
phidata
python-dotenv
yfinance
duckduckgo-search
google-cloud-aiplatform
